Remove all Webpack files and dependencies from app
Webpack is installed with the rails generators but is not needed at
this moment in time.
It is currently causing CI to fail when it comes to running the tests.
So as we do not need it for this project, I have removed it to try and
fix this issue.