Refactor pages_controller and the new and edit forms
Edit and New actions are now rendering the _form partial. This is to
ensure my code is DRY.
The form_with declaration is slightly different for the original new
action form than the _form partial, though they are doing the same
thing. This has changed the wording to the submit button for the
rendered new action form, which is why the spec wording has now changed
for the click_button part.