Reformat html and styling throughout the app
The original html layout and styling has needed to be changed to ensure
the sidebar can sit properly on the index page.
Also the original implementation was done with little understanding.
Now that my knowledge of html and css has increased, I can ensure the
layout fits in better with conventions.